cleaned up childproc and project1.c

+++ project1/child_proc.c
@@ -32,7 +32,7 @@ int main(int argc, char *argv[]) {
     */
 
     int* matrix = shmat(shmid_matrix, NULL, 0);
-    int* partial_sum = shmat(shmid_partial_sum, NULL, 0);
+    long* partial_sum = shmat(shmid_partial_sum, NULL, 0);
    
     long sum = sum_matrix(start_row_id, end_row_id, matrix);
     partial_sum[partial_sum_index] = sum;
